Lib reorganization: separate node.h file, only one compiled lib
graph.h should only handle external functions. Create files node.h and
node.c for all functions and struct definitions related to Node.

Reduce libraries to only libgraph.la
